+++ crates/rmf_site_editor/src/interaction/anchor.rs
@@ -17,8 +17,7 @@
 
 use crate::{
     interaction::{gizmo::Gizmo, IntersectGroundPlaneParams, *},
-    keyboard::DebugMode,
-    site::{Anchor, Category, Delete, Dependents, SiteAssets, Subordinate},
+    site::{Anchor, Category, Delete, Dependents, SiteAssets, Subordinate}, DebugMode,
 };
 use bevy::{ecs::hierarchy::ChildOf, prelude::*};
 use rmf_site_picking::{Hovered, Selectable, Selected};
