+++ crates/rmf_site_editor/src/keyboard.rs
@@ -16,37 +16,22 @@
 */
 
 use crate::{
-    site::{AlignSiteDrawings, Delete},
-    CreateNewWorkspace, CurrentWorkspace, WorkspaceLoader, WorkspaceSaver,
+    site::{AlignSiteDrawings, Delete}, CreateNewWorkspace, CurrentWorkspace, DebugMode, WorkspaceLoader, WorkspaceSaver
 };
 use bevy::{prelude::*, window::PrimaryWindow};
 use bevy_egui::EguiContexts;
 use bevy_impulse::*;
 use rmf_site_camera::resources::ProjectionMode;
-use rmf_site_picking::{KeyboardServices, Selection};
+use rmf_site_picking::Selection;
 
-#[derive(Debug, Clone, Copy, Resource)]
-pub struct DebugMode(pub bool);
+/// plugin for managing input settings for rmf_site_editor
+pub struct EditorInputPlugin;
 
-impl FromWorld for DebugMode {
-    fn from_world(_: &mut World) -> Self {
-        DebugMode(false)
-    }
-}
-
-pub struct KeyboardInputPlugin;
-
-impl Plugin for KeyboardInputPlugin {
+impl Plugin for EditorInputPlugin {
     fn build(&self, app: &mut App) {
         app.init_resource::<DebugMode>()
             .add_systems(Last, handle_keyboard_input);
 
-        let keyboard_just_pressed =
-            app.spawn_continuous_service(Last, keyboard_just_pressed_stream);
-
-        app.insert_resource(KeyboardServices {
-            keyboard_just_pressed,
-        });
     }
 }
 
@@ -151,8 +136,3 @@ pub fn keyboard_just_pressed_stream(
         orders.for_each(|order| order.streams().send(StreamOf(*key_code)));
     }
 }
-
-// #[derive(Resource)]
-// pub struct KeyboardServices {
-//     pub keyboard_just_pressed: Service<(), (), StreamOf<KeyCode>>,
-// }
